+++ src/Dictionary.php
@@ -53,7 +53,7 @@ final class Dictionary implements MemberOrderedMap
         return match (true) {
             $member instanceof ParameterAccess &&
             ($member instanceof MemberList || $member instanceof ValueAccess) => $member,
-            is_iterable($member) => InnerList::from(...$member),
+            is_iterable($member) => InnerList::new(...$member),
             default => Item::fromAssociative($member),
         };
     }